requirements.txt
aiohttp==1.3.3
aiopg==0.13.0
appdirs==1.4.2
async-timeout==1.1.0
chardet==2.3.0
click==6.7
Flask==0.12
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.5
lxml==3.7.3
MarkupSafe==0.23
multidict==2.1.4
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pkg-resources==0.0.0
psycopg2==2.7.1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
trafaret==0.9.0
trafaret-config==1.0.1
tushare==0.7.4
Werkzeug==0.11.15
yarl==0.9.8
